pyproject.toml
[project]
name = "frappe-bench"
description = "CLI to manage Multi-tenant deployments for Frappe apps"
readme = "README.md"
license = "GPL-3.0-only"
requires-python = ">=3.7"
authors = [
    { name = "Frappe Technologies Pvt Ltd", email = "developers@frappe.io" },
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Natural Language :: English",
    "Operating System :: MacOS",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: User Interfaces",
    "Topic :: System :: Installation/Setup",
]
dependencies = [
    "Click>=7.0",
    "GitPython~=2.1.15",
    "honcho",
    "Jinja2~=3.0.3",
    "python-crontab~=2.6.0",
    "requests",
    "semantic-version~=2.8.2",
    "setuptools>40.9.0",
    "tomli;python_version<'3.11'",
]
dynamic = [
    "version",
]

[project.scripts]
bench = "bench.cli:cli"

[project.urls]
Changelog = "https://github.com/frappe/bench/releases"
Documentation = "https://frappeframework.com/docs/user/en/bench"
Homepage = "https://frappe.io/bench"
Source = "https://github.com/frappe/bench"

[build-system]
requires = [
    "hatchling>=1.6.0",
]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "bench/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
  "/bench"
]

[tool.hatch.build.targets.wheel]
include = [
  "/bench"
]
